Simplify image validation by separating responsibilities
