+++ src/Core/Function.php
@@ -7,7 +7,7 @@ require "Database.php";
 function dd($value)
 {
     echo "<pre>";
-     var_dump($value);
+    var_dump($value);
     echo "</pre>";
     die();
 }
