+++ src/Views/Template/Categories/create.phtml
@@ -33,6 +33,7 @@ require base_path("Views/Partials/head.phtml") ?>
 
                 <!-- Upload new image -->
                 <input type="file" name="image"
+                    accept=".png, .jpeg, .jpg"
                     class="mt-2 w-full text-sm text-gray-700 
                file:mr-4 file:py-2 file:px-4 file:rounded-lg file:border-0 
                file:text-sm file:font-semibold
