+++ src/Views/Template/home.phtml
@@ -2,8 +2,8 @@
 <html lang="en">
 <?php
 require base_path("Views/Partials/head.phtml");
- ?>
- 
+?>
+
 <body>
     <?php require base_path("Views/Partials/nav.phtml") ?>
     <?php require base_path("Views/Partials/header.phtml") ?>
@@ -13,23 +13,25 @@ require base_path("Views/Partials/head.phtml");
             <h2 id="clothies" class="text-center font-bold text-5xl text-indigo-600"><?= $category["categoryName"] ?></h2>
             <div class="mx-auto max-w-2xl px-4 py-16 sm:px-6 sm:py-24 lg:max-w-7xl lg:px-8">
                 <div class="grid grid-cols-1 gap-x-6 gap-y-10 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 xl:gap-x-8">
-                    <?php foreach ($products[$category['id']] as $product) : ?>
-                        <a href="#" class="group">
-                            <img src="/assets/images/Products/<?= $product["productImage"] ?>" alt="Tall slender porcelain bottle with natural clay textured body and cork stopper." class="aspect-square w-full rounded-lg bg-gray-200 object-cover group-hover:opacity-75 xl:aspect-7/8" />
-                            <div class="flex justify-between items-center">
-                                <h3 class="mt-1 text-sm text-gray-700"><?= $product["productName"] ?></h3>
-                                <p class="mt-1 text-lg font-medium text-gray-900">$<?= $product["productPrice"] ?></p>
-                            </div>
-                            <div class="mt-2 flex items-center gap-2">
-                                <button class="flex-1 bg-indigo-600 hover:bg-blue-700 text-white text-sm font-medium py-1.5 px-3 rounded-lg">Add to Cart</button>
-                                <button class="text-red-500 hover:text-red-600">
-                                    <svg class="h-6 w-6" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24">
-                                        <path stroke-linecap="round" stroke-linejoin="round" d="M4.318 6.318a4.5 4.5 0 016.364 0L12 7.636l1.318-1.318a4.5 4.5 0 116.364 6.364L12 21.364l-7.682-7.682a4.5 4.5 0 010-6.364z" />
-                                    </svg>
-                                </button>
-                            </div>
-                        </a>
-                    <?php endforeach; ?>
+                    <?php if (isset($products[$category['id']])) : ?>
+                        <?php foreach ($products[$category['id']] as $product) : ?>
+                            <a href="#" class="group">
+                                <img src="/assets/images/Products/<?= $product["productImage"] ?>" alt="Tall slender porcelain bottle with natural clay textured body and cork stopper." class="aspect-square w-full rounded-lg bg-gray-200 object-cover group-hover:opacity-75 xl:aspect-7/8" />
+                                <div class="flex justify-between items-center">
+                                    <h3 class="mt-1 text-sm text-gray-700"><?= $product["productName"] ?></h3>
+                                    <p class="mt-1 text-lg font-medium text-gray-900">$<?= $product["productPrice"] ?></p>
+                                </div>
+                                <div class="mt-2 flex items-center gap-2">
+                                    <button class="flex-1 bg-indigo-600 hover:bg-blue-700 text-white text-sm font-medium py-1.5 px-3 rounded-lg">Add to Cart</button>
+                                    <button class="text-red-500 hover:text-red-600">
+                                        <svg class="h-6 w-6" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24">
+                                            <path stroke-linecap="round" stroke-linejoin="round" d="M4.318 6.318a4.5 4.5 0 016.364 0L12 7.636l1.318-1.318a4.5 4.5 0 116.364 6.364L12 21.364l-7.682-7.682a4.5 4.5 0 010-6.364z" />
+                                        </svg>
+                                    </button>
+                                </div>
+                            </a>
+                        <?php endforeach; ?>
+                    <?php endif; ?>
                 </div>
             </div>
     </div>
